Refactor from tss-react to scss
Removed tss-react and installed scss to minimize client-side rendering and avoid duplication of processing.

While reading through tss-docs, I realized that a confusingly-worded section indicated that while SSR was supported, it means that the components are going to be rendered on the server side and the client side. To avoid duplication of processing and remove server processing time, tss-react was changed for scss